requirements.txt
Django==2.1
django-widget-tweaks==1.4.2
djangorestframework==3.8.2
Markdown==2.6.11
mysqlclient==1.3.13
pytz==2018.5
